Complete the refactor of util.
